fix: partly work around paper PlayerProfile issues
this basically applies two things to the paper player profile resolver:
- it inserts a random generated name in case the name wasn't completed by the profile resolver
- it throws a better exception in case the player uuid didn't get resolved
This is happening due to a paper issue: calling complete() on a player profile does not actually complete the stuff like name and uuid, it only completes the properties (but the information is present). This is not an issue in the spigot resolver.